% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ir_tables.R
\name{combine_ir_tables}
\alias{combine_ir_tables}
\title{Makes a single IR table from multiple IR devices}
\usage{
combine_ir_tables(path_to_parsed_ir)
}
\arguments{
\item{path_to_parsed_ir}{Full path to the directory containing
parsed IR csv files.}
}
\value{
A dataframe of IR counter data.
}
\description{
Combines a directory of multiple parsed IR files into a single df with
a complete timeseries and NAs counts converted to zeros.
}
