% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_zero_counts.R
\name{insert_zero_counts}
\alias{insert_zero_counts}
\title{Replaces NA counts with zeros.}
\usage{
insert_zero_counts(df)
}
\arguments{
\item{df}{A dataframe of IR counts.}
}
\value{
A dataframe of IR data with missing values filled in.
}
\description{
Replaces NA counts with zeros.
}
\examples{
df <- tibble::tibble(
PlacementID = c(102, 103, 104, 105),
date = c("19-06-04", "19-06-04", "19-07-04", "19-07-06"),
time = c("14:00", "14:00", "13:01", "12:02"),
count = c(15, NA, 12, NA))
insert_zero_counts(df)
}
