% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_complete_timeseries.R
\name{make_complete_timeseries}
\alias{make_complete_timeseries}
\title{Creates complete timeseries of IR data.}
\usage{
make_complete_timeseries(df)
}
\arguments{
\item{df}{A dataframe of IR counter data.}
}
\value{
A dataframe with any missing dates inserted.
}
\description{
Inserts new rows into timeseries for dates that are inside days of operation.
}
\examples{
df <- tibble::tibble(
PlacementID = c(102, 102, 102, 102),
date = c("19-06-04", "19-06-07", "19-06-08", "19-06-10"),
time = c("14:00", "14:00", "13:01", "12:02"),
count = c(15, 13, 12, 50))
make_complete_timeseries(df)
}
