% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ir_table.R
\name{make_ir_table}
\alias{make_ir_table}
\title{Processes a single parsed IR file}
\usage{
make_ir_table(df)
}
\arguments{
\item{df}{The dataframe of IR counts from a single device.}
}
\value{
A dataframe of IR count data with complete dates and any missing counts
converted to zeros.
}
\description{
Creates a complete timeseries for a single parsed IR file and converts NAs
counts to zeros.
}
\examples{
df <- tibble::tibble(
PlacementID = c(102, 102, 102, 102),
date = c("19-06-04", "19-06-07", "19-06-08", "19-06-10"),
time = c("14:00", "14:00", "13:01", "12:02"),
count = c(15, 13, 12, 50))
make_ir_table(df)
}
