% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ir.R
\name{parse_ir}
\alias{parse_ir}
\title{Parses a raw IR txt file into a csv.}
\usage{
parse_ir(ir_file, ir_dir, parsed_dir, dump_dir)
}
\arguments{
\item{ir_file}{Name of raw IR file (do not include the path to file).}

\item{ir_dir}{Path to the directory containing raw IR file.}

\item{parsed_dir}{Path to directory where parsed csv should be written.}

\item{dump_dir}{Path to directory where potentially corrupt files
should be written.}
}
\value{
A parsed csv file.
}
\description{
Parses a raw IR txt file into a csv.
}
